Add basic JUnit testing with 3 test cases
Rename Test.java to DebugTests.java to avoid naming conflict, add unit
testing cases for BasicTrajectory velocity/acceleration,
TankDriveTrajectory velocity and WaypointEx at the beginning and end of
a path.
